Move code for register-searching pass into singledispatch.py
Move the code for the pass that looks for singledispatch functions and
corresponding register calls into singledispatch.py.

This also adds some documentation about why we need a separate pass for
looking for singledispatch functions instead of just adding this
searching to one of the already existing passes.

This also fixes some circular imports caused by moving functions into
singledispatch.py.
