+++ mypyc/irbuild/main.py
@@ -30,13 +30,14 @@ from mypy.build import Graph
 
 from mypyc.common import TOP_LEVEL_NAME
 from mypyc.errors import Errors
+from mypyc.irbuild.singledispatch import find_singledispatch_register_impls
 from mypyc.options import CompilerOptions
 from mypyc.ir.rtypes import none_rprimitive
 from mypyc.ir.module_ir import ModuleIR, ModuleIRs
 from mypyc.ir.func_ir import FuncIR, FuncDecl, FuncSignature
 from mypyc.irbuild.prebuildvisitor import PreBuildVisitor
 from mypyc.irbuild.vtable import compute_vtable
-from mypyc.irbuild.prepare import build_type_map, find_singledispatch_register_impls
+from mypyc.irbuild.prepare import build_type_map
 from mypyc.irbuild.builder import IRBuilder
 from mypyc.irbuild.visitor import IRBuilderVisitor
 from mypyc.irbuild.mapper import Mapper
